--- a/executor/src/errors.rs
+++ b/executor/src/errors.rs
@@ -4,7 +4,6 @@ use config::WorkerId;
 use std::fmt::Debug;
 use store::StoreError;
 use thiserror::Error;
-use types::CertificateDigest;
 
 #[macro_export]
 macro_rules! bail {
@@ -49,9 +48,6 @@ pub enum SubscriberError {
 
     #[error("Storage failure: {0}")]
     StoreError(#[from] StoreError),
-
-    #[error("Error occurred while retrieving certificate {0} payload: {1}")]
-    PayloadRetrieveError(CertificateDigest, String),
 
     #[error("Consensus referenced unexpected worker id {0}")]
     UnexpectedWorkerId(WorkerId),