--- a/primary/src/lib.rs
+++ b/primary/src/lib.rs
@@ -40,7 +40,7 @@ pub use crate::{
         responses::{CertificatesResponse, PayloadAvailabilityResponse},
         BlockHeader,
     },
-    block_waiter::{BlockCommand, BlockWaiter, GetBlockResponse},
+    block_waiter::{BlockCommand, BlockWaiter},
     grpc_server::metrics::EndpointMetrics,
     metrics::PrimaryChannelMetrics,
     primary::{NetworkModel, PayloadToken, Primary, PrimaryWorkerMessage, CHANNEL_CAPACITY},