--- a/test_utils/src/lib.rs
+++ b/test_utils/src/lib.rs
@@ -94,20 +94,6 @@ macro_rules! test_new_certificates_channel {
     };
 }
 
-#[macro_export]
-macro_rules! test_get_block_commands {
-    ($e:expr) => {
-        types::metered_channel::channel(
-            $e,
-            &prometheus::IntGauge::new(
-                primary::PrimaryChannelMetrics::NAME_GET_BLOCK_COMMANDS,
-                primary::PrimaryChannelMetrics::DESC_GET_BLOCK_COMMANDS,
-            )
-            .unwrap(),
-        );
-    };
-}
-
 ////////////////////////////////////////////////////////////////
 /// Keys, Committee
 ////////////////////////////////////////////////////////////////