This means to start a discussion re: reverting #842 and #738 (with the former reverted for building on top of the later)

On the benchmarks, in CI, this shows a 64% increase in throughput and 60% decrease in latency:
- the base of this PR (main) shows [17K TPS at 1.6s latency](https://github.com/MystenLabs/narwhal/commit/e2cae6382c0b988195990d3102c549dfdbdd6a30#commitcomment-83224735)
- the tip of this PR shows [28K TPS at 1.0s latency](https://github.com/MystenLabs/narwhal/commit/293cecd0b997df937ad141f7e70f0f4b7bf5b2e8#commitcomment-83225149)

At home, on a (voluntarily) puny laptop (which will do much better on e.g. serde):
- the base of this PR (main) shows [42.9K TPS at 709ms latency](https://gist.github.com/huitseeker/31963368870f5d707b48fb5dd745e05d)
- the tip of this PR shows [45.4K TPS at 500ms latency](https://gist.github.com/huitseeker/733b621ef40886ce080b9499f8ae22e8)

It also seems this might improve some long-running integration tests:
- [before](https://gist.github.com/huitseeker/7dac7a894cd93054a06af513b00cb6ea)
- [after](https://gist.github.com/huitseeker/0aea1bb43dc8b14f7380869939ff93d9)

This fixes #858, and might improve #841.