@asonnino 2022-09-07
Are we re-introducing the batch-loader instead of the batch waiter?