@akichidis 2022-09-12
Regarding the batch_loader and the performance changes (@huitseeker we talked a little bit offline about those today) . A couple of info regarding that which unfortunately didn't express more vocally on my PR #738 (added in description but probably got lost there). 

Executor [was configured](https://github.com/MystenLabs/narwhal/commit/babd3d8ee6278652a8fee3b41764787203c17cc1#diff-c1fa00b27f461be84b2e8f9f93974c238229bb7a87964886c0fe316ed825f4e0L299) to use the [batch_store](https://github.com/MystenLabs/narwhal/blob/ef92fc0d455494e7b97aaa02a528356f45169a24/node/src/lib.rs#L47) which as part of #738 is changed to use a separate store (the `temp_batch_store` was introduced) to keep the batches fetched from the workers. The `batch_store` is the exact same one that [a worker node is using as well](https://github.com/MystenLabs/narwhal/blob/ef92fc0d455494e7b97aaa02a528356f45169a24/node/src/lib.rs#L393). In a SUI deployment when we initialise the Narwhal's storage we do it only once and we pass the same `NodeStorage` instance to both the `worker` and the `primary` nodes. I see that currently we use the [NodeRestarter](https://github.com/MystenLabs/sui/blob/dced2e77ac9f93e66311ba9bd536e820d868696e/crates/sui-core/src/authority_server.rs#L267) - before the nodes were initialised manually - but the underneath practice is the same, NodeRestarter will pass [the same store instance to both nodes](https://github.com/MystenLabs/narwhal/blob/ef92fc0d455494e7b97aaa02a528356f45169a24/node/src/restarter.rs#L84). So practically both the `primary` and `worker` would share the same storage. The worker would use it to store in the `batch_store` its (and others) batches and the primary would use the same `batch_store`, via the Executor, to effectively store the "fetched" batches from the worker.

The [Subscriber](https://github.com/MystenLabs/narwhal/commit/babd3d8ee6278652a8fee3b41764787203c17cc1#diff-dd570646b526219ac76c68c9546526c43df1bf27085a0949d1355a15bd5d84b8L65) was implemented to first check in the `batch_store` about the required certificate's batches. If all existed (`notify_read` etc) then the [certificate would be delivered](https://github.com/MystenLabs/narwhal/commit/babd3d8ee6278652a8fee3b41764787203c17cc1#diff-dd570646b526219ac76c68c9546526c43df1bf27085a0949d1355a15bd5d84b8L65) for execution. Now, given that the `Subscriber` would look directly to the worker's store (batch_store) my understanding is that we would never observe any sort of network issues - or at least get affected by them as we practically never leverage the network communication for that purpose. Also the `batch_loader` wouldn't really do much , apart from fetching via network the batch from the worker and "rewriting" the same data again to the same (shared) store - `batch_store`. So if we revert the PRs lets keep in mind that we'll re-introduce the above behaviour - both primary & worker use the exact same store (`batch_store`) to keep their batches. Although it will make things definitely more robust (as we are talking now about a monolith) it will give us the false impression about the primary <--> worker communication when we request the batches which I would strongly recommend then refactor asap.