@huitseeker 2022-09-12
Yep,I think the overall outcome we want to get to is a better story for the `BlockWaiter` in the presence of large messages, i.e. an Executor-focused variant of https://github.com/MystenLabs/narwhal/issues/396. I no longer think this PR, which was there for expediency anyway, is the best way to address this.